requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1

# Google / GCP
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.127.0
google-cloud-firestore>=2.16.0
google-cloud-pubsub>=2.21.0

# LLM
google-generativeai>=0.5.0

# Telegram
python-telegram-bot>=21.0

# Dashboard
streamlit>=1.33.0

# Utilities
httpx>=0.27.0
tenacity>=8.3.0
